Bump dependencies & revamp tests
