+++ package.json
@@ -11,7 +11,7 @@
     "type": "git",
     "url": "https://github.com/ecies/js.git"
   },
-  "version": "0.4.5",
+  "version": "0.4.6",
   "engines": {
     "node": ">=16.0.0"
   },
@@ -35,14 +35,14 @@
     "test": "jest"
   },
   "dependencies": {
-    "@noble/ciphers": "^0.3.0",
+    "@noble/ciphers": "^0.4.0",
     "@noble/curves": "^1.2.0",
     "@noble/hashes": "^1.3.2"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.5",
-    "@types/node": "^20.8.2",
-    "@types/node-fetch": "^2.6.6",
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.9.0",
+    "@types/node-fetch": "^2.6.9",
     "https-proxy-agent": "^7.0.2",
     "jest": "^29.7.0",
     "node-fetch": "^2.7.0",
