+++ src/utils/symmetric.ts
@@ -6,7 +6,7 @@ import { sha256 } from "@noble/hashes/sha256";
 
 import { symmetricAlgorithm, symmetricNonceLength } from "../config";
 import { AEAD_TAG_LENGTH, XCHACHA20_NONCE_LENGTH } from "../consts";
-import { aes256gcm } from "./compat";
+import { aes256cbc, aes256gcm } from "./compat";
 
 export function aesEncrypt(key: Uint8Array, plainText: Uint8Array): Uint8Array {
   // TODO: Rename to symEncrypt
@@ -19,7 +19,7 @@ export function aesDecrypt(key: Uint8Array, cipherText: Uint8Array): Uint8Array 
 }
 
 export function deriveKey(master: Uint8Array): Uint8Array {
-  // 32 bytes shared secret for aes and xchacha20
+  // 32 bytes shared secret for aes256 and xchacha20 derived from HKDF-SHA256
   return hkdf(sha256, master, undefined, undefined, 32);
 }
 
@@ -30,6 +30,9 @@ function _exec(is_encryption: boolean, key: Uint8Array, data: Uint8Array): Uint8
     return callback(aes256gcm, key, data, symmetricNonceLength());
   } else if (algorithm === "xchacha20") {
     return callback(xchacha20, key, data, XCHACHA20_NONCE_LENGTH);
+  } else if (algorithm === "aes-256-cbc") {
+    // aes-256-cbc is always 16 bytes iv and there is no AEAD tag
+    return callback(aes256cbc, key, data, 16, 0);
   } else {
     throw new Error("Not implemented");
   }
@@ -39,14 +42,15 @@ function _encrypt(
   func: (key: Uint8Array, nonce: Uint8Array, AAD?: Uint8Array) => Cipher,
   key: Uint8Array,
   plainText: Uint8Array,
-  nonceLength: number
+  nonceLength: 12 | 16 | 24,
+  tagLength: 16 | 0 = AEAD_TAG_LENGTH
 ): Uint8Array {
   const nonce = randomBytes(nonceLength);
   const cipher = func(key, nonce);
   const ciphered = cipher.encrypt(plainText); // TAG + encrypted
 
-  const encrypted = ciphered.subarray(0, ciphered.length - AEAD_TAG_LENGTH);
-  const tag = ciphered.subarray(-AEAD_TAG_LENGTH);
+  const encrypted = ciphered.subarray(0, ciphered.length - tagLength);
+  const tag = ciphered.subarray(ciphered.length - tagLength);
   return concatBytes(nonce, tag, encrypted);
 }
 
@@ -54,9 +58,10 @@ function _decrypt(
   func: (key: Uint8Array, nonce: Uint8Array, AAD?: Uint8Array) => Cipher,
   key: Uint8Array,
   cipherText: Uint8Array,
-  nonceLength: number
+  nonceLength: 12 | 16 | 24,
+  tagLength: 16 | 0 = AEAD_TAG_LENGTH
 ): Uint8Array {
-  const nonceTagLength = nonceLength + AEAD_TAG_LENGTH;
+  const nonceTagLength = nonceLength + tagLength;
   const nonce = cipherText.subarray(0, nonceLength);
   const tag = cipherText.subarray(nonceLength, nonceTagLength);
   const encrypted = cipherText.subarray(nonceTagLength);
